chore: add phpcs and fix files
